requirements.txt
openai-whisper
numpy
torch
torchaudio
ffmpeg-python